refactor(server): extract workspace target launch
