+++ server/proliferate/server/cloud/workspaces/target_launch/__init__.py
@@ -1,0 +1,1 @@
+"""Desktop target-launch workspace flow."""
